backend done

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({user: req.user.id})
 
   res.status(200).json(goals)
 })
@@ -14,7 +15,8 @@ const addGoals = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id,
   })
 
   res.status(200).json(goal)
@@ -25,7 +27,21 @@ const editGoal = asyncHandler(async (req, res) => {
   
   if(!goal) {
     res.status(400)
-    throw new Error('Goal not find')
+    throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  //check if user exist
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //chekc if logged in user matches goal user
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -38,7 +54,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
   
   if(!goal) {
     res.status(400)
-    throw new Error('Goal not find')
+    throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  //check if user exist
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //chekc if logged in user matches goal user
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.findByIdAndDelete(req.params.id)
